# interpreter: removes unneeded stack walks